Fixes page refreshes when using back and forward buttons
The onpopstate method tests whether the non-firefox browser sent a
pop state event contains the same URL as the page that it was
transitioning to. The check is not only done to see if event is not
null since the pop state event will be fired by the browser only for
pushState calls and therefore as long as we are on the Proteus page
the event will be valid. This allows the correct segment to get loaded
instead of a full page refresh.

Fixes: Defect 18351