Added correct parameter for tracking original_document_url
The original_document_url meant to be initialized on the first
page load of the application was changing and therefore created
a page refresh on subsequent use of the back and forward buttons.
This change ensures that the original url stays the same for the
specific application and SPA protocol is followed.