History logic modified to check each page instead of landing page
Modified the utils functionality to create a history event for every page
in the application (including the most recent page load). This is due to
the fact that the first solution verified against the the original URL that
was used when first navigating to the page. This check worked as long as
you stayed on our application (the one using utils.js). Once you navigate
away and then click back or even refresh our page, the new landing URL would
be set as the one to verify against, but all of the history events contained
the old landing URL and thus fail.

This required modification of utils.js to verify the URL of the
page you are going back or forward to against the URL placed in the
history event. This meant I had to modify the onclick event handler and
the accept_url function to do a replacestate on each page load. The
onpopstate event had to be modified to verify the page that the user was
navigating to instead of the self.original_document_url to be consistent.
This was tested by navigating around the SPA that uses utils.js as well as
navigating to other SPAs such as the cloudsupplements example in
lda-examples. Refreshing the page or navigating back to another SPA and
back to our page worked correctly.